requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchmetrics>=0.11.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.3.0
scikit-image>=0.18.0
PySide6>=6.0.0
Pillow>=8.0.0
tqdm>=4.60.0
pyyaml>=5.4.0
tensorboard>=2.7.0
pytest>=6.0.0
albumentations>=1.0.0